Disable -b/--notify pty regress test for now (re: 667034ff)
The referenced commit made -b/--notify much better at notifying
multiple terminating background jobs, but it still occasionally
misses one if multiple jobs terminate all at once. This race
condition will hopefully be worked out at some point, but it's a
low-priority problem as the notification will appear at the next
prompt in any case, and we still have many worse bugs to fix.

But the associated pty regression test is failing fairly often on
the github runners in the meantime, so this disables it for now.