Workaround to allow compiling on NetBSD/{sparc,alpha}
At least on sparc, frexpl(3) does not seem to be included in
libm.so as expected:

  $ uname -a
  NetBSD netsparc.inlv.org 9.3 NetBSD 9.3 (GENERIC) #0:
  Thu Aug  4 15:30:37 UTC 2022
  mkrepro@mkrepro.NetBSD.org:/usr/src/sys/arch/sparc/compile/GENERIC
  sparc
  $ strings -a /usr/lib/libm.so|grep frexp
  frexpf
  frexp
  $NetBSD: s_frexpl.c,v 1.3 2013/02/12 21:40:19 martin Exp $
  $NetBSD: s_frexpf.c,v 1.10 2007/08/21 20:12:27 drochner Exp $
  $NetBSD: s_frexp.c,v 1.13 2008/09/28 18:54:55 christos Exp $
  frexp
  frexpf
  $

So, frexpl is not found as a symbol name using strings(1), though
the corresponding .c file is apparently compiled in. The generic
iffe lib test for it also fails to detect it. As a result,
_lib_frexpl is not defined, so this prototype is declared:

src/lib/libast/features/float:
880: #if !_lib_frexpl || _npt_frexpl
881: 	printf("extern _ast_fltmax_t	frexpl(_ast_fltmax_t, int*);\n");
882: #endif

Since _ast_fltmax_t is defined as 'double' on this system (which is
the same size as 'long double'), this prototype is incompatible
with the standard 'long double' prototype declared in NetBSD's
system header /usr/include/math.h, and the compilation fails.

This build failure has only been found on NetBSD/sparc and
NetBSD/alpha. On NetBSD/x86_64, 'strings -a' does find frexpl in
/usr/lib/libm.so.

src/lib/libast/features/float:
- To avoid the problem, check for frexpl(3) by compiling a test
  program instead of searching for it in the libraries with a
  generic iffe test. This also adds a basic verification that it
  works as expected.
- For consistency, do the same for ldexpl(3), even though it does
  not have this problem on NetBSD/sparc.

src/lib/libast/comp/frexpl.c,
src/lib/libast/sfio/sfhdr.h:
- Use 'long double' instead of '_ast_fltmax_t'. As of a34e8319, we
  require C90 which specifies 'long double', so there is no more
  need for that K&R C compatibility workaround.

Thanks to @bremensaki for the report.
Resolves: https://github.com/ksh93/ksh/issues/528