optget: fix out-of-bounds read in label()
The label() function outputs labels like the bolded 'version' label
shown when you do somecommand --version. In that specific case, a
literal "version" label string is passed instead of a label from an
option/self-doc string. label() checks for a '+' preceding that
label, which in that case causes the function to read one byte
before the literal string as the 's' pointer is never increased.
The fix is to check that the pointer has been increased first.

Related: https://github.com/att/ast/issues/836
Resolves: https://github.com/ksh93/ksh/pull/541