Bring in-word completion in line with the book (#580)
In both editing modes, using = or * completion from within a word
fails to replace the word; any trailing characters are printed at
the end of the completion. In addition, no characters under or
beyond the cursor (beyond the cursor in vi mode) are taken into
account when constructing the list of options. As an example, in
emacs mode with the cursor on i, $ ls foi<^M-=> can produce the
following completion menu:

1) foible
2) foil
3) foo.bar

According to Bolsky and Korn, pp. 105-106 and 114-115, completion
characters are meant to be appended to the word. This implies that
the completion operation takes place on the entire word, as it does
in pdksh, oksh and mksh. In the example above, full-word completion
would result in a menu of:

1) foible
2) foil

src/cmd/ksh93/edit/vi.c:
- Save cur_virt and increment until either virtual[cur_virt] or
  virtual[cur_virt+1] is non-alphanumeric. Move the cursor back to
  the original position if completion fails or in case of a
  completion menu.

src/cmd/ksh93/edit/emacs.c:
- Save cur and increment until out[cur] is non-alphanumeric. Move
  the cursor back to the original position if completion fails or
  in case of = completion with a menu, no count set and = as the
  most-recently entered key.