cleanup: rm JOBS symbol, POSIX signal ifdefs, waitpid(2) fallback
jobs.h defines the JOBS symbol that is set if the system is deemed
to support jobs control. This symbol was set if either SIGCHLD or
SIGCLD is a signal on the current system. There is no POSIX
compliant system made in recent decades that doesn't have SIGCHLD,
whereas SIGCLD is completely obsolete.

As expected, bit rot set in long ago; a test build of ksh with JOBS
disabled failed. It is not worth fixing.

All changed files:
- Remove JOBS symbol and corresponding fallbacks.
- Remove preprocessor checks for the following signal symbols that
  have been specified by POSIX for decaces: SIGCHLD, SIGHUP,
  SIGPIPE, SIGQUIT, SIGSTOP, SIGTSTP, SIGTTIN, SIGTTOU.
- Remove all checks for and references to the obsolete SIGCLD
  symbol (all systems have been using SIGCHLD instead for decades).

src/lib/libast/comp/waitpid.c:
- Removed. All POSIX-ish operating systems provide waitpid(2), so
  none of the fallback code has been used for many years.