cleanup: package: remove SHELLMAGIC test for hashbang path
This code, which was added in 2002, only added a hashbang path to
shell scripts when running on Cygwin (tested by looking for
/bin/env.exe) or EMX (tested by looking for /emx/bin/sh.exe).

Cygwin has not required the use of a hashbang path for at least a
decade and a half (if it ever did). Absent a hashbang path, /bin/sh
is used to run the shell script, which is also what this code does.

As for EMX, its last stable release was in 1998. Nuff said.
https://en.wikipedia.org/wiki/EMX_(programming_environment)