Tweak the code style to approach 1990s standard
I've always preferred to keep the historic character of this code
base, but some things have finally gotten on my nerves. There is
still not going to be a comprehensive reformatting, but now that we
require at least a C90 compiler and no longer pretend to support
C++ (see a34e8319), I think it's time to eliminate a few constructs
that may have been needed to placate C++ or ancient K&R compilers.

First of all, null pointers were expressed in several ways:
- NIL(type), e.g., NIL(char*)
- (type)0, e.g., (char*)0
- NiL
- 0
- NULL, rarely

The inconsistency of this was quite annoying. This commit replaces
NIL(type), NiL, and (type)0 with NULL. Plain '0' for null pointers
is not yet replaced because that needs to be carefully fixed on a
case by case basis.

Second, I'm removing the obsolete register keyword. I've tried to
fix the tab formatting mess that ensued in variable declarations.

Finally, this commit changes all instances of
	return(foo);
to
	return foo;