read.c: fix 'read -r -d' for !SHOPT_MULTIBYTE (re: 11e528b7)
The GitHub CI run showed that 'read -r -d DELIM', for values of
DELIM other than a newline, was broken in the regression test run
with SHOPTs disabled. The regression test added in the referenced
commit exposed that breakage.

The bug is in an optimisation code block for 'read -r' that only
gets compiled in if SHOPT_MULTIBYTE is disabled.

I was tempted to simply delete that block, but the fix is trivial:
when stripping trailing delimiters, check for the current delimiter
(delim) instead of '\n'.