cleanup: remove !_PACKAGE_ast dead code
Some of the libast and ksh code theoretically supports building
without (the rest of) libast by including fallback code that is
compiled if _PACKAGE_ast is not defined. But all the compile
commands in the Mamfiles pass -D_PACKAGE_ast, so all that is dead
code and can be removed, along with the symbol itself[*].

Since we've never used or tested any of that, our changes probably
broke it long ago anyway, even if it wasn't broken before.

Defining _PACKAGE_ast also force-defined _hdr_locale for ksh93 (via
features/locale), regardless of whether locale.h is on the system.
The '#ifdef _hdr_locale' checks are hereby removed from sh/init.c.
________
[*] The only exception is src/cmd/INIT/mamake.c, which is built
before libast, but has optional _PACKAGE_ast code to enable AST
option parsing and self-documentation. And that does work. Try:
   $ bin/package make
   $ bin/package use
   $ cc -O -D _PACKAGE_ast -I $INSTALLROOT/include/ast \
	-L $INSTALLROOT/lib -l ast -l iconv \
	-o $INSTALLROOT/bin/mamake src/cmd/INIT/mamake.c
   $ mamake --man