Re-fix assoc array compound assignment syntax check (re: 9c037b04)
Looking for "]=" in the literal argument string is not correct: it
doesn't take account of shell grammar. So the fix left bugs like:

    $ ksh -c 'fn=([foo_key]=foo_val [bar_key\]=])'
    ksh: [bar_key]=]: not found
    $ ksh -c 'badword=([x]=1 \[badword]=good)'
    ksh: [badword]=good: not found

The 'not found' means it tried to execute these as commands (if an
executable named '[badword]=good' were on $PATH, it would be run).

But the lexer has already done all the necessary analysis and added
an ARG_ASSIGN flag to the argument node if it is a syntactically
valid shell assignment (in this context, including one of the form
[foo]=bar).

The parser checked for that flag on the element of the compound
assignment, but failed to check it for subsequent arguments (which
is why the original bug only occurred in the second element and up
of a compound assignment). So all we need to do is add the missing
check (and remove my broken one).

After this commit:

    $ ksh -c 'fn=([foo_key]=foo_val [bar_key\]=])'
    ksh: syntax error at line 1: `[bar_key\]=]' unexpected
    $ ksh -c 'badword=([x]=1 \[badword]=good)'
    ksh: syntax error at line 1: `[badword]=good' unexpected

...which is as it should be.

Related: https://github.com/ksh93/ksh/issues/427