tput testing adjustments (#630)
Currently, if tput testing reveals support for neither termcap nor
terminfo, _pth_tput is disabled in edit.c, but the shell attempts
multi-line editing anyway if multiline is set to on. If this
happens, the following errors can occur:

  * During the course of typing:
      - Navigating leftwards across the window edge echoes the
        first line again, with second line material in place of PS1
      - Navigating rightwards across the window edge produces a new
        line, and second line characters are revealed sequentially
  * When reached from history:
      - All leftward navigation is ineffectual
      - vi mode: the cursor appears on the last line after a
        PS1-sized offset
      - Rightward navigation reveals the command sequentially
      - Inserting text echoes the command line again after every
        character

This causes problems on FreeBSD in particular because features/cmds
uses the ansi terminal to check for termcap support, and FreeBSD
has an especially restrictive termcap entry for ansi. Moreover, the
script checks for sgr0 (terminfo) and me (termcap) capabilities,
neither of which is actually used in edit.c or emacs.c. The issue
can be resolved by using the vt520 terminal for testing instead and
removing the unneeded checks.

  * features/cmds: Use the vt520 terminal in preference to ansi for
    tput testing, and remove the checks for the unused sgr0 and me
    capabilities.
  * edit.c: Turn off multiline functionality if tput support is not
    available (i.e., there is no tput in the path, or neither
    termcap nor terminfo is supported on the system).

These changes have been tested on the following operating systems,
and all of them have a specified vt520 terminal:

  * Linux
  * macOS
  * FreeBSD
  * OpenBSD
  * NetBSD
  * OmniOS (illumos-based)
  * Haiku