edit.c: Don't insert space and backspace after each underscore (#627)
On macOS and Haiku, entering an underscore (_) at or shortly
preceding the window's edge with multiline enabled causes the
character before the underscore to be replaced onscreen. When
reaching the command through history navigation, other graphical
errors occur, including but possibly not limited to:

  * The start of the command line overlaps the prompt
  * When entering text at the beginning of the command line, it
    appears only one character at a time while the remainder of the
    command line stays in place

The apparent cause is at edit.c:1025, legacy code that inserts a
space and \b when an underscore is entered. This code has been in
the source since 1986 at the latest, and its presence is not
mentioned in any of the RELEASE documents from that time. Although
its purpose is unclear, it is conceivable that it may be an ancient
workaround for a single-byte text encoding scheme.

src/cmd/ksh93/edit/edit.c: ed_putchar():
- Remove this long-obsolete workaround.

Resolves: https://github.com/ksh93/ksh/issues/620