tests/pty.sh: increase delays on some tests (re: 79a908f4, a35a47b8)
As of 79a908f4, ksh appropriately invokes tput(1) to get the escape
sequences necessary for multiline editing, even on FreeBSD. But on
a Raspberry Pi running FreeBSD arm64, a number of pty tests started
failing as a result. None of those failures occurred on a faster
machine running the same OS, so it's evidently a timing issue
pushed over the edge by invoking tput(1). So this commit increases
the delays on the tests that failed on the slow machine from 15 to
20 ms.

Resolves: https://github.com/ksh93/ksh/issues/634