xec.c: allow for slightly higher function call depth
In xec.c, context checkpoint buffers (which are large-ish) are
allocated on the AST stack using st(a)kalloc instead of declaring
them as non-pointer local variables, to avoid a C stack overflow
on a large function recursion depth. This was a change made in
ksh 93u+ 2011-12-24.

There is still some room to move some more large local variables
to the AST stack, which makes a segfault on a high function
recursion depth even less likely. This commit does that.

For clarity, this commit also stop using a local stkp variable
for the shell default stack point (sh.stk) and just references
it directly.

This commit also corrects the type of set_instance() to long; it
returns a long value (nelem) that should not be converted to int.