Revert expanding extended patterns from unquoted vars (re: aab40092)
Turns out that bug 3 in the referenced commit is actually intended
behaviour. It doesn't seem to be documented anywhere, but it makes
sense: for compatibility with the Bourne/POSIX shells, unquoted
variables should only expand basic shell patterns, as parentheses
are regular characters in that context. OTOH, when extended shell
patterrns (with parentheses) are expanded when given as literal
strings, they are a straightforward syntax error in POSIX, so only
in that case do they not pose a POSIX compatibility problem.

Extended patterns (with parentheses) are disabled by internally
backslash-escaping S_ESC (\) and S_EPAT (&|()) characters here:

2429:	char *sp = "&|()";
2430:	while(c = *sp++)
2431:	{
2432:		if(state[c]==0)
2433:			state[c] = S_EPAT;
2434:	}
[...]
2454:	if(n==S_ESC || n==S_EPAT)
2455:	{
2456:		/* don't allow extended patterns in this case */
2457:		mp->patfound = mp->pattern;
2458:		sfputc(stkp,ESCAPE);
2459:	}

Meanwhile, ksh *does* expand brace expressions from unquoted
variables in non-POSIX mode (see 009445bf), which is also
incompatible with POSIX, so I fixed that for the POSIX mode. But in
ksh mode, the behaviour is currently inconsistent: why expand
braces from unquoted variables but not extended patterns?

Be that as it may, the behaviour has been this way since 1995 and
changing it will probably break some script or other, so at least
it should not change on the 1.0 branch. As for 1.1 and/or later, we
will have to consider whether to change this at some point.

src/cmd/ksh93/sh/macro.c:
- Revert as per above.

src/cmd/ksh93/sh.1:
- Fix the documentation to describe the actual behaviour.