Fix signal exit status with EXIT trap set
Reproducer script:

    trap 'echo hoi' EXIT
    sleep 3
    echo end

Run the script, press Ctrl+C within 3 seconds, and examine the exit
status (echo $?). The exit status will be 2. Expected is 2+256=258
(on ksh93; on most shells it's 2+128=130, on yash it's 2+384=386).

Analysis: the EXIT trap is run from sh_done() right before the
shell terminates. This involves running a command, which causes
sh_exec() to reset sh.lastsig, which is then used in sh_done() to
determine if the received signal should be resent to self to
generate the correct signal exit status.

src/cmd/ksh93/sh/fault.c: sh_done():
- Save sh.lastsig on entry.
- After potentially running the EXIT trap, use the saved value
  instead of sh.lastsig to decide whether to reissue the signal.